Refactored code and files
