--- lib/dev.rb
@@ -1,1 +1,0 @@
-puts 'hello from indent.rb'
