+++ style.css
@@ -19,6 +19,7 @@ p {
  }
 #bob {
   font-size: 12em;
+  funky
 }
    
 div {
