--- test.rb
@@ -1,10 +1,0 @@
-require 'json'
-
-# css_props_file = File.read('./assets/css-properties.json')
-# css_props = JSON.parse(css_props_file)
-
-# p css_props['property']
-
-p "Hello  there mr!".count(' ')
-
-p "Ru uu  by".index(' ')
